[util] Add useTheme hook and refactor components
